requirements.txt
Flask
python-dotenv
openai==1.51.0
flask-cors
requests